Factor out types from `assistants` module in `types`
