+++ async-openai/src/types/assistant.rs
