+++ async-openai/src/types/assistant_file.rs
