--- async-openai/src/types/assistants/mod.rs
@@ -1,16 +1,0 @@
-mod assistant;
-mod assistant_file;
-mod assistant_impls;
-mod message;
-mod message_file;
-mod run;
-mod step;
-mod thread;
-
-pub use assistant::*;
-pub use assistant_file::*;
-pub use message::*;
-pub use message_file::*;
-pub use run::*;
-pub use step::*;
-pub use thread::*;
