+++ async-openai/src/types/message_file.rs
