+++ async-openai/src/types/mod.rs
@@ -1,9 +1,24 @@
 //! Types used in OpenAI API requests and responses.
 //! These types are created from component schemas in the [OpenAPI spec](https://github.com/openai/openai-openapi)
-mod assistants;
+mod assistant;
+mod assistant_file;
+mod assistant_impls;
+mod message;
+mod message_file;
+mod run;
+mod step;
+mod thread;
+
+pub use assistant::*;
+pub use assistant_file::*;
+pub use message::*;
+pub use message_file::*;
+pub use run::*;
+pub use step::*;
+pub use thread::*;
+
 mod impls;
 mod types;
-pub use assistants::*;
 use derive_builder::UninitializedFieldError;
 pub use types::*;
